Move the `proppy` module under the `@bayou` namespace.
