+++ local-modules/@bayou/proppy/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "proppy",
+  "name": "@bayou/proppy",
   "version": "1.0.0",
 
   "dependencies": {
