+++ local-modules/env-server/package.json
@@ -6,7 +6,7 @@
     "is-running": "^2.1.0",
 
     "hooks-server": "local",
-    "proppy": "local",
+    "@bayou/proppy": "local",
     "@bayou/see-all": "local",
     "@bayou/util-common": "local"
   }
